+++ src/main/kotlin/GetOptError.kt
@@ -1,0 +1,5 @@
+package com.hankadler.io
+
+import kotlin.Exception
+
+class GetOptError(message: String): Exception(message)